branch-4.0: handle schemas without delete sign

+++ be/src/service/point_query_executor.cpp
@@ -95,7 +95,9 @@ static void get_missing_and_include_cids(const TabletSchema& schema,
         missing_cids.insert(slot->col_unique_id());
     }
     // insert delete sign column id
-    missing_cids.insert(schema.columns()[schema.delete_sign_idx()]->unique_id());
+    if (schema.delete_sign_idx() >= 0) {
+        missing_cids.insert(schema.columns()[schema.delete_sign_idx()]->unique_id());
+    }
     if (target_rs_column_id == -1) {
         // no row store columns
         return;
